Incremental processing of alignments by using parsimony to insert sequences into an existing tree. 1. PhyloTree::initializeAllPartialPars now returns int (the number of    parsimony memory blocks that it used up) (in tree/phylotree.cpp) 2. PhyloTree::addTaxonMP drafted (which decides, for a internal    node and a leaf node, that could be inserted into a parsimony    tree, which of the possible insertion locations would result    in maximum parsimony), in tree/placement.cpp. 3. PhyloTree::addNewTaxaToTree now    (a) adds nodes by maximum parsimony (I have in mind allowing        adding nodes by maximum likelihood, too, but I don't have        that working as yet).    (b) attempts to place new nodes in batches (via B clause in        incremental_mode). Default batch size: 1.  0 means all.    (c) can insert a *subset* of the nodes in each batch (via I clause        in incremental_mode). Default insertion count: 1. 0 means all. 4. Added convert_int_nothrow in utils/tools.h (and .cpp).    It's called during processing of Params::incremental_method.
Notes:
A. Because parsimony and maximum likelihood scoring modifies data
stored (in the NeighbourVector of each PhyloNode), to parallelise
the searching (for the nodes in a batch?), it might be
necessary to make a copy of the tree for each thread (ouch!).
I haven't written any of the code to do *that* yet.  I'm not even
sure that will be necessary (but it might be, for parsimony
scoring).  It mightn't be necessary.  It might be possible to
tweak parsimony calculations (if the interface uses the neighbour
information on three nodes, rather than the nodes themselves, you
could use it to answer "what if?" Questions: given what A sees of B,
and B sees of A, and what a node C sees of a new leaf node, X,
what would the cost be if C were linked in between A and B?).
Such questions wouldn't (shouldn't) interfere with each other
at all! If the Cs and Xs don't overlap. (Once all the parsimony
calculations have been done in both directions for every pair
of linked nodes, in the existing tree).  Linear scaling?

B. When adding a node by maximum likelihood, the likelihood (of the
tree) depends, in part, on the lengths of the edges leading from
the new internal node (which are, I think, calculated, but aren't -
at the moment - available) (I may need to add more & (reference)
parameters to addTaxonML).  It'd be better if that information
were available.

incremental_method clauses:
Bn How many nodes to batch at a time (default 1) (0 means ALL)
Ha What heuristic (a==algorithm) to use to narrow the search
   (Not implemented).
Cf What cost function to use (f=={MP} or {ML} for now) (default {MP})
   (Only {MP} works as yet).
In How many nodes to insert at a time (0 means ALL in batch)
   (Can't be more than the number to batch) (default 1).
La What clean-up methods to use after each insert (not implemented)
Aa What clean-up methods to use after each batch (not implemented)
Ta What clean-up methods to use after all nodes added (not implemented)

default:
-incremental B1+C{MP}+I1

useful for testing; set the #define of BODGE_ONE_NODE_IN_TEN to 1,
and pass in:
-incremental B0+C{MP}+I0