Parsimony Placement now parallelises parsimony calculations "across each level of the inputs", via ParallelParsimonyCalculator. TargetBranch::computeState uses an instance of it. ParallelParsimonyCalculator's approach is 1. for each parsimony branch score that is wanted at this level    (sequentially); determine what its direct inputs are;    to be calculated at the next level. 2. Recursively calculate all the scores at the next level. 3. For each parsimony branch score that is wanted at this level    (In parallel); determine it from its direct inputs.
This is simple, but it's not ideal (as there's a "join" at each
"level" of processing, and "levels" are calculated working down
from the "top" level (when it would be slightly better if they were
calculated looking "upward" from the bottom level, since the
work would "thin out" only at the top, rather than at both the
top and the bottom).  In large trees, though, the difference
shouldn't matter too much.

The approach used in ParallelParsimonyCalculator is adequate
for parsimony calculation but it mightn't be appropriate for
likelihood (as likelihood vectors are larger!).   For parallelising
likelihood calculation "across the tree" (rather than across sites),
a "countdown latch" model might work better.