Added code to decenttree.cpp for doing at least some processing of alignments in .fasta files: 1. loadAlignmentIntoDistanceMatrix    (And supporting functions, checkLastTwoSequenceLengths     And processSequenceLine) 2. loadSequenceDistancesIntoMatrix    (And supporting function, correctDistance, and a     Supporting class, SiteInfo), which    makes use of vectorHammingDistance and    countBitsSetInEither). 3. Added a -fasta parameter (in main()).    Made -dist a synonym for -in (if you pass -dist, and -fasta    you'll (eventually) get a distance file *output* rather    than an input).    (Though, I haven't added the code for writing .dist files yet).
I added all this is I could benchmark inputs that are too large for
IQTree to handle well (on my work laptop).  Note that the distance
calculation in decenttree takes rather longer than that in IQTree
(Because, for example, non-constant sites that are not parsimony-
informative *are* counted, by decenttree) (for a ~17K-sequence
input that one difference is enough to double the running time
for distance matrix calculation!). The difference will depend on
the input, but is typically between 2:1 and 3:1.